refactor: Simplify SQL parser utility functions and improve readability
